+++ component/frontend/Model/DownloadIDLabels.php
@@ -10,7 +10,9 @@ namespace Akeeba\ReleaseSystem\Site\Model;
 defined('_JEXEC') or die();
 
 use Akeeba\ReleaseSystem\Admin\Model\DownloadIDLabels as AdminDownloadIDLabels;
+use Akeeba\ReleaseSystem\Admin\Model\Mixin\ClearCacheAfterActions;
 
 class DownloadIDLabels extends AdminDownloadIDLabels
 {
+	use ClearCacheAfterActions;
 }